login ui completed

+++ app/src/main/java/com/hemp/works/login/data/model/Doctor.kt
@@ -3,8 +3,8 @@ package com.hemp.works.login.data.model
 data class Doctor(
     var profile     : String? = null,
     var status      : String,
-    var credit      : Int,
-    var cart        : Int,
+    var credit      : Long,
+    var cart        : Long,
     var name        : String,
     var clinic      : String,
     var mobile      : String,
